+++ src/widgets/mod.rs
@@ -1,7 +1,9 @@
+pub mod log;
 pub mod simple_list;
 pub mod simple_text;
 pub mod text_fragments;
 
+pub use log::Log;
 pub use simple_list::SimpleList;
 pub use simple_text::SimpleText;
 pub use text_fragments::TextFragments;